requirements.txt
Django==5.0.6
easyocr==1.7.1
numpy==1.26.4
Pillow==10.2.0
pdf2image==1.17.0
djangorestframework==3.15.1
python-dateutil==2.9.0.post0